+++ organization_service/app/schemas/storage_distance.py
@@ -12,7 +12,7 @@ class StorageDistanceCopyBaseSchema(BaseModel):
 
     storage_id: int
     organisation_id: int
-    distance: float
+    distance: int
 
 
 class StorageDistanceCopySchema(StorageDistanceCopyBaseSchema):
